sp_panic_handler

Struct AbortGuard

pub struct AbortGuard { /* private fields */ }

RAII guard for whether panics in the current thread should unwind or abort.

Sets a thread-local abort flag on construction and reverts to the previous setting when dropped. Does not implement Send on purpose.

Note: Because we restore the previous value when dropped, you are encouraged to leave the AbortGuard on the stack and let it destroy itself naturally.

Implementations

impl AbortGuard

pub fn force_unwind() -> AbortGuard

Create a new guard. While the guard is alive, panics that happen in the current thread will unwind the stack (unless another guard is created afterwards).

pub fn force_abort() -> AbortGuard

Create a new guard. While the guard is alive, panics that happen in the current thread will abort the process (unless another guard is created afterwards).

pub fn never_abort() -> AbortGuard

Create a new guard. While the guard is alive, panics that happen in the current thread will never abort the process (even if AbortGuard::force_abort() guard will be created afterwards).